Remove all extraneous entries from .gitignore and commit meta files
Previous .gitignore file contained a number of bad ignores. Most notably,
it ignored .meta files, which should *never* be omitted. The meta files
should now be properly synchronised.

Beware of extraneous files when commiting. If you see a file, a folder, or
a file type that should be ignored, please add it back to .gitignore.